Divided the test class into two classes to test the email class
